+++ lib/hexapdf/type/font_type0.rb
@@ -54,7 +54,9 @@ module HexaPDF
 
       # Returns the CID font of this type 0 font.
       def descendant_font
-        document.cache(@data, :descendant_font) { document.deref(self[:DescendantFonts][0]) }
+        document.cache(@data, :descendant_font) do
+          document.wrap(document.deref(self[:DescendantFonts][0]))
+        end
       end
 
       # Returns the writing mode which is either :horizontal or :vertical.